requirements.txt
libnacl
discord.py==2.3.2